+++ CHANGELOG.md
@@ -21,10 +21,11 @@
 - [761](https://github.com/vegaprotocol/data-node/issues/761) - Delete all badger stores, `SQL` stores only from now
 - [566](https://github.com/vegaprotocol/data-node/issues/566) - Liquidity provision pagination 
 
-
 ### 🐛 Fixes
 - [705](https://github.com/vegaprotocol/data-node/issues/705) - Market Depth returning incorrect book state
-- [730](https://github.com/vegaprotocol/data-node/issues/730) - Event bus subscriptions with party and market filter not working 
+- [730](https://github.com/vegaprotocol/data-node/issues/730) - Event bus subscriptions with party and market filter not working
+- [678](https://github.com/vegaprotocol/data-node/issues/678) - Add new trading mode variant
+- [776](https://github.com/vegaprotocol/data-node/issues/776) - Add support for missing proposal errors
 
 ## 0.52.0